# openapi/karma-v1.yaml
openapi: 3.1.0
info:
  title: Karma Ecosystem API
  version: 1.0.0
  summary: Payment intent, evidence verification, and risk alert APIs.
servers:
  - url: http://127.0.0.1:8811
security:
  - bearerAuth: []
paths:
...
  /v1/settlement/{taskId}/partial:
    post:
      summary: Partial settlement by value percent
      operationId: partialSettlement
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialSettlementRequest'
      responses:
        '200':
          description: Updated settlement state
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/settlement/{taskId}/regret:
    post:
      summary: Buyer regret settlement
      operationId: regretSettlement
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegretRequest'
      responses:
        '200':
          description: Updated settlement state
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/settlement/{taskId}/dispute:
    post:
      summary: Open dispute for settlement task
      operationId: openDispute
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisputeRequest'
      responses:
        '200':
          description: Updated settlement state
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/settlement/{taskId}/auto-arbitrate:
    post:
      summary: Run automatic arbitration rules
      operationId: autoArbitrate
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated settlement state after automatic arbitration
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PartialSettlementRequest:
      type: object
      properties:
        settled_value_percent:
          type: number
          exclusiveMinimum: 0
          maximum: 100
        reason:
          type: string
      required: [settled_value_percent]
    RegretRequest:
      type: object
      properties:
        buyer_identity_id:
          type: string
        reason:
          type: string
    DisputeRequest:
      type: object
      properties:
        reason:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
          required: [code, message, requestId]
      required: [error]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'